refactory moving common code to wrapper
